Update operation scheduling for SWAT+
Operation scheduling is extended to the use with SWAT+ now. Minor
things are not solved yet (e.g. calculating HRU fractions because PHU
information is not in the mgt input table anymore). Using phu_fr in
SWAT+ now will result in an error when scheduling.
Next steps: fixing end of year and skip year