Fix data handling when no initial management.sch
The SWAT+Editor can write projects with no initial management.sch.
Also the pointer in the file.cio is set null then.
The farmR did not consider this case and an error occured as the
initial management.sch was not found.
This commit now consideres the case that no initial management.sch
is available in reading writing and resetting the input files.
Also now the file.cio is always written, to avoid a null pointer.